chore: cleanup a mess and duplicates
