+++ typescript/packages/common-system/src/index.ts
@@ -1,3 +1,4 @@
 export { h, Fragment } from "./jsx.js";
 export * from "./adapter.js";
 export * from "./rule-builder.js";
+export * from "./view.js";
